normalize errors and fix lint
