+++ packages/datadog-plugin-jest/src/index.js
@@ -38,23 +38,6 @@ function getTestSpanMetadata (tracer, test) {
   }
 }
 
-// based on https://github.com/facebook/jest/blob/main/packages/jest-circus/src/formatNodeAssertErrors.ts#L41
-function formatJestError (errors) {
-  let error
-  if (Array.isArray(errors)) {
-    const [originalError, asyncError] = errors
-    if (originalError === null || !originalError.stack) {
-      error = asyncError
-      error.message = originalError
-    } else {
-      error = originalError
-    }
-  } else {
-    error = errors
-  }
-  return error
-}
-
 class JestPlugin extends Plugin {
   static get name () {
     return 'jest'
@@ -84,9 +67,8 @@ class JestPlugin extends Plugin {
       this.tracer._exporter._writer.flush()
     })
 
-    this.addSub('ci:jest:test:err', (errors) => {
-      if (errors) {
-        const error = formatJestError(errors)
+    this.addSub('ci:jest:test:err', (error) => {
+      if (error) {
         const span = storage.getStore().span
         span.setTag(TEST_STATUS, 'fail')
         span.setTag('error', error)
